// src/app/ficha-de-matricula/models/daclaracion-de-compromiso.ts
export class DeclaracionDeCompromiso {
    id:              number=0;
    descripcion:     string='';
    vigente:         boolean=false;
    fechaDeRegistro: string='';
    colegioId:       number=0;
}
export class FilaDeclaracionDeCompromiso {
    nro:number=0;
    id:              number=0;
    descripcion:     string='';
    vigente:         boolean=false;
    fechaDeRegistro: string='';
    colegioId:       number=0;
}